+++ pyobjc-core/Modules/objc/helpers.h
@@ -8,6 +8,7 @@ extern int PyObjC_setup_nscoder(void);
 extern int PyObjC_setup_nsdecimal(PyObject* m);
 extern int PyObjC_setup_nsobject(void);
 extern int PyObjC_setup_simd(void);
+extern int PyObjC_setup_nsinvocation(void);
 
 extern PyObject* _Nullable pythonify_nsdecimal(const void* value);
 extern int depythonify_nsdecimal(PyObject* value, void* out);